Add frame grabbing timeout handling for simple camera controller
Typically something like this is handled the the device finding, open, close, level which polls at slow rate. In this driver the highest level of abstraction is camera. For now we are just hacking this in, instead of re-architecting the driver. IF we use this driver more in the future it would make sense to add a polled device finding/open/close loop.